requirements.txt
torch>=2.1
torchaudio>=2.1
pandas~=2.2.3
numpy~=2.0.2
pyyaml~=6.0.2
librosa
torchvision~=0.22.0